rename scorecard route
